+++ lib/puppetlabs_spec_helper/puppetlabs_spec/puppet_internals.rb
@@ -1,7 +1,3 @@
-# Initialize puppet for testing by loading the
-# 'puppetlabs_spec_helper/puppet_spec_helper' library
-require 'puppetlabs_spec_helper/puppet_spec_helper'
-
 module PuppetlabsSpecHelper
   module PuppetInternals
     # parser_scope is intended to return a Puppet::Parser::Scope
