+++ puppetlabs_spec_helper.gemspec
@@ -17,6 +17,8 @@ Gem::Specification.new do |s|
   s.add_dependency "rspec", "~> 2.9.0"
   s.add_dependency "rspec-puppet", "~> 1.0.1"
   s.add_dependency "puppet-lint", "~> 0.3.2"
+  s.add_dependency "puppet", "~> 3.5.1"
+  s.add_dependency "mocha", "~> 0.14.0"
 
   s.files        = Dir.glob("lib/**/*") + %w(LICENSE CHANGELOG)
   s.require_path = 'lib'
