--- spec/lib/puppet/type/spechelper.rb
@@ -1,6 +1,0 @@
-require 'puppet/type'
-
-Puppet::Type.newtype(:spechelper) do
-  @doc = "This is the spechelper type"
-end
-
