+++ spec/spec_helper.rb
@@ -1,0 +1,2 @@
+require 'puppetlabs_spec_helper'
+# Dir['./spec/support/**/*.rb'].sort.each { |f| require f }
